requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
orjson==3.10.7
pydantic-settings==2.6.1
python-multipart==0.0.9
motor==3.6.0
bcrypt==4.2.0
PyJWT==2.9.0
cloudinary==1.41.0
stripe==10.10.0
httpx==0.27.2

email-validator==2.1.1
